Refactor: Styled 컴포넌트 분리
